admin
This reverts commit 00a0b5fb5e72f75d32107fb91ffb95e069cbaddf, reversing
changes made to 76a12d9363eb0f5848460076b6d14db066b7e477.

+++ src/main/java/BookStore/Model/Sale.java
@@ -3,12 +3,12 @@ package BookStore.Model;
 public class Sale {
     private int id;
     private String percent;
-    private double priceSale;
+    private int priceSale;
 
     public Sale() {
     }
 
-    public Sale(int id, String percent, double priceSale) {
+    public Sale(int id, String percent, int priceSale) {
         this.id = id;
         this.percent = percent;
         this.priceSale = priceSale;
@@ -30,11 +30,11 @@ public class Sale {
         this.percent = percent;
     }
 
-    public double getPriceSale() {
+    public int getPriceSale() {
         return priceSale;
     }
 
-    public void setPriceSale(double priceSale) {
+    public void setPriceSale(int priceSale) {
         this.priceSale = priceSale;
     }
